refactor: Adde security to the swagger.json to allow API Key input for docs look up, changed token required decorator to the decorators folder, added the flask email new package to the requirements.txt

+++ alembic/versions/8c0527cff320_token_schema.py
@@ -9,7 +9,7 @@ from typing import Sequence, Union
 
 from alembic import op
 import sqlalchemy as sa
-
+from sqlalchemy.engine.reflection import Inspector
 
 # revision identifiers, used by Alembic.
 revision: str = '8c0527cff320'
@@ -17,11 +17,21 @@ down_revision: Union[str, None] = 'fedadcb66914'
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
 
+def column_exists(table_name, column_name):
+    inspector = Inspector.from_engine(op.get_bind())
+    columns = inspector.get_columns(table_name)
+    return any(column['name'] == column_name for column in columns)
 
 def upgrade():
-    op.add_column('user_table', sa.Column('birth_date', sa.Date(), nullable=True))
-    op.add_column('admins', sa.Column('auth_token', sa.String(), nullable=True))
+    if not column_exists('user_table', 'birth_date'):
+        op.add_column('user_table', sa.Column('birth_date', sa.Date(), nullable=True))
+    
+    if not column_exists('admins', 'auth_token'):
+        op.add_column('admins', sa.Column('auth_token', sa.String(), nullable=True))
 
 def downgrade():
-    op.drop_column('user_table', 'birth_date')
-    op.drop_column('admins', 'auth_token')
+    if column_exists('user_table', 'birth_date'):
+        op.drop_column('user_table', 'birth_date')
+    
+    if column_exists('admins', 'auth_token'):
+        op.drop_column('admins', 'auth_token')
